# json/todo/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Ica MinButik - Todo",
    "version" : "1.0.0",
    "title" : "ToDo API",
    "contact" : {
      "email" : "goran.roseen@acando.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/ica-egenkontroll/todo/1.0.0",
...
  "schemes" : [ "https" ],
  "paths" : {
    "/todos" : {
      "get" : {
        "tags" : [ "clients" ],
        "summary" : "searches for todos",
        "description" : "By passing in the appropriate options, you can search for\navailable todos in the system\n",
        "operationId" : "searchTodo",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "$top",
          "in" : "query",
          "description" : "Show only the first n items, see [OData Paging - Top](http://docs.oasis-open.org/odata/odata/v4.0/odata-v4.0-part1-protocol.html#_Toc445374630)",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "$skip",
          "in" : "query",
          "description" : "Skip the first n items, see [OData Paging - Skip](http://docs.oasis-open.org/odata/odata/v4.0/odata-v4.0-part1-protocol.html#_Toc445374631)",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "$filter",
          "in" : "query",
          "description" : "Filter items by property values, see [OData Filtering](http://docs.oasis-open.org/odata/odata/v4.0/odata-v4.0-part1-protocol.html#_Toc445374625), note that expand is not supported",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "$select",
          "in" : "query",
          "description" : "Select properties to be returned, see [OData Select](http://docs.oasis-open.org/odata/odata/v4.0/odata-v4.0-part1-protocol.html#_Toc445374620), note that expand is not supported",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/TodoItem"
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          }
        }
      }
    },
    "/todos/{id}" : {
      "get" : {
        "tags" : [ "clients" ],
        "summary" : "Find Todo by Id",
        "description" : "Find a Todo by Id",
        "operationId" : "getTodoById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Id of the Todo that needs to be fetched",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/TodoItem"
            }
          },
          "400" : {
            "description" : "error"
          },
          "404" : {
            "description" : "Could not find an entry with the given id."
          }
        }
      },
      "delete" : {
        "tags" : [ "admins" ],
        "summary" : "Delete ToDo by Id",
        "description" : "Deletes a Todo item in the system",
        "operationId" : "deleteTodo",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "ID of the Todo that needs to be deleted",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Delete successful"
          },
          "400" : {
            "description" : "error"
          },
          "404" : {
            "description" : "Could not find an entry with the given id."
          }
        }
      }
    },
...
  },
  "definitions" : {
    "TodoItem" : {
      "required" : [ "title", "todoType" ],
      "properties" : {
        "id" : {
          "type" : "string",
          "example" : "123456"
        },
        "title" : {
          "type" : "string",
          "example" : "Städa charken"
        },
        "todoType" : {
          "type" : "string",
          "example" : "basic",
          "description" : "Must be one of \"basic\", \"measure\", \"yesno\""
        },
        "note" : {
          "type" : "string",
          "example" : "Gör så här"
        },
        "contentUrl" : {
          "type" : "string",
          "format" : "url",
          "example" : "https://episerver.ica.se/getContent?id=4711"
        },
        "createdAt" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "2018-03-14 15:22"
        },
        "dueAt" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "2018-04-12 12:00"
        },
        "assignedToId" : {
          "type" : "string",
          "example" : "d290f1ee-6c54-4b01-90e6-d701748f0851"
        },
        "assignedToName" : {
          "type" : "string",
          "example" : "Frukt & Grönt"
        },
        "assignedToObjectType" : {
          "type" : "string",
          "example" : "Group"
        },
        "isCompleted" : {
          "type" : "boolean",
          "example" : true
        },
        "completedBy" : {
          "$ref" : "#/definitions/User"
        },
        "completedAt" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "2018-04-12 12:00"
        },
        "priority" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 1
        },
        "categoryId" : {
          "type" : "string",
          "example" : "123"
        },
        "categoryTitle" : {
          "type" : "string",
          "example" : "HACCP"
        },
        "tags" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Tag"
          }
        },
        "modifiedOn" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "2018-03-14 15:22"
        }
      },
      "example" : {
        "note" : "Gör så här",
        "completedAt" : "2018-04-12 12:00",
        "todoType" : "basic",
        "title" : "Städa charken",
        "priority" : 1,
        "assignedToId" : "d290f1ee-6c54-4b01-90e6-d701748f0851",
        "assignedToName" : "Frukt & Grönt",
        "assignedToObjectType" : "Group",
        "tags" : [ {
          "modifiedOn" : "2018-03-14 15:22",
          "sortId" : 1,
          "id" : "123456",
          "title" : "Rengöring"
        }, {
          "modifiedOn" : "2018-03-14 15:22",
          "sortId" : 1,
          "id" : "123456",
          "title" : "Rengöring"
        } ],
        "dueAt" : "2018-04-12 12:00",
        "createdAt" : "2018-03-14 15:22",
        "contentUrl" : "https://episerver.ica.se/getContent?id=4711",
        "modifiedOn" : "2018-03-14 15:22",
        "id" : "123456",
        "categoryId" : "123",
        "categoryTitle" : "HACCP",
        "completedBy" : {
          "id" : "d290f1ee-6c54-4b01-90e6-d701748f0851",
          "fullname" : "Göran Roseen"
        },
        "isCompleted" : true
      }
    },
    "Tag" : {
      "type" : "object",
      "required" : [ "title" ],
      "properties" : {
        "id" : {
          "type" : "string",
          "example" : "123456"
        },
        "title" : {
          "type" : "string",
          "example" : "Rengöring"
        },
        "sortId" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 1
        },
        "modifiedOn" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "2018-03-14 15:22"
        }
      },
      "example" : {
        "modifiedOn" : "2018-03-14 15:22",
        "sortId" : 1,
        "id" : "123456",
        "title" : "Rengöring"
      }
    },
...
    "User" : {
      "required" : [ "fullname", "id" ],
      "properties" : {
        "id" : {
          "type" : "string",
          "example" : "d290f1ee-6c54-4b01-90e6-d701748f0851",
          "description" : "id is the user identity from SIA"
        },
        "fullname" : {
          "type" : "string",
          "example" : "Göran Roseen"
        }
      },
      "example" : {
        "id" : "d290f1ee-6c54-4b01-90e6-d701748f0851",
        "fullname" : "Göran Roseen"
      }
    },
...
  },
...
}